examples/agent_demo/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
azure-cognitiveservices-speech==1.25.0
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
dataclasses-json==0.5.7
decorator==5.1.1
faiss-cpu==1.7.3
fastapi==0.92.0
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.16.3
google-cloud-speech==2.17.3
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
gTTS==2.3.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
langchain==0.0.146
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.2
openai==0.27.2
openapi-schema-pydantic==1.2.4
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
ply==3.11
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pyflakes==3.0.1
PyJWT==2.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.7.1
PyYAML==6.0
redis==4.5.3
regex==2023.3.23
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.25.0
tenacity==8.2.2
tomli==2.0.1
tqdm==4.65.0
twilio==7.17.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
vocode==0.1.99
websockets==10.4
yarl==1.8.2
